python: Remove miscellaneous uses of boost
These include:

- boost::tuple -- replaced with std::tuple
- boost::align -- replaced with alignas and std::align

- boost::implicit_cast
  The few uses seemed unnecessary. One use was to cast to void*,
  which should always be valid and could just be a static_cast.
  The other was to make the implicit conversion of numeric
  arguments when calling Python API explicit, which doesn't
  seem useful.

- boost::compressed_pair
  This was used in one place to hold a function object and
  associated call policies, which may be an empty class. It
  was replaced with a bespoke object that applies the EBCO
  only to the latter.

(Internal change: 2341795)
(Internal change: 2341823)
(Internal change: 2341829)

+++ pxr/external/boost/python/converter/builtin_converters.hpp
@@ -20,7 +20,6 @@
 # include "pxr/external/boost/python/detail/none.hpp"
 # include "pxr/external/boost/python/handle.hpp"
 # include "pxr/external/boost/python/ssize_t.hpp"
-# include <boost/implicit_cast.hpp>
 # include <string>
 # include <complex>
 # include <boost/limits.hpp>
@@ -167,15 +166,15 @@ PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(unsigned PXR_BOOST_PYTHON_LONG_LONG, ::PyLon
 #if PY_VERSION_HEX >= 0x03000000
 PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(char, converter::do_return_to_python(x), &PyUnicode_Type)
 PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(char const*, converter::do_return_to_python(x), &PyUnicode_Type)
-PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(std::string, ::PyUnicode_FromStringAndSize(x.data(),implicit_cast<ssize_t>(x.size())), &PyUnicode_Type)
+PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(std::string, ::PyUnicode_FromStringAndSize(x.data(),x.size()), &PyUnicode_Type)
 #else
 PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(char, converter::do_return_to_python(x), &PyString_Type)
 PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(char const*, converter::do_return_to_python(x), &PyString_Type)
-PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(std::string, ::PyString_FromStringAndSize(x.data(),implicit_cast<ssize_t>(x.size())), &PyString_Type)
+PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(std::string, ::PyString_FromStringAndSize(x.data(),x.size()), &PyString_Type)
 #endif
 
 #if defined(Py_USING_UNICODE)
-PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(std::wstring, ::PyUnicode_FromWideChar(x.data(),implicit_cast<ssize_t>(x.size())), &PyUnicode_Type)
+PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(std::wstring, ::PyUnicode_FromWideChar(x.data(),x.size()), &PyUnicode_Type)
 # endif 
 PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(float, ::PyFloat_FromDouble(x), &PyFloat_Type)
 PXR_BOOST_PYTHON_TO_PYTHON_BY_VALUE(double, ::PyFloat_FromDouble(x), &PyFloat_Type)
